Solr-based browsing, closes #272
The browsing of User Things (Objects, Collections and Stories) is now done by fetching data from SOLR.
Additionally fields are created at indexing time, which will later need to be reviewed (some fields are duplicates of the ones for MDRs).
Improvements to the query service make it possible to do advanced visibility search, so everything is indexed but only shown to the correct people.