Separating themes configuration, introducing info.conf for graphic themes
Themes are now separated: the general configuration of the theme, with SOLR keys etc. still lives in the PortalTheme, as well as the theme name and subdomain configuration. The "display theme" configuration has now moved to info.conf files that are in the theme directory (in 'themes/default/info.conf' for example for the default theme). These info.conf files can override keys of the default theme.

The general configuration of themes live in one or more '_themes.yml' files in 'conf/'.