Displaying MDRs directly via mongo, and more
At indexing time we now save the mapping that is produced to mongo (due to the limitations of Salat we need to persist it as a MongoDBObject that we then transform). A special MetadataAccessor implementation takes care of giving access to the common fields.

Now both user objects and institutional objects that are linked in are displayed on the collection edition page.