Bookmarks
Each registered user now gets a default Bookmarks collection. This collection can't be deleted; its name and description and visibility are fixed. We pass the name and description to the Bookmarks collection ListItem (in list view), full view (in the collection view) and edit view via the controller, and these are fetched via the i18n messages.