Re-viewing escaping and encoding for SOLR fields
Always XML-encoding field values before parsing them. Do a check on URLs, in case we get invalid URLs (this should be checked by the mapping engine / the schema validation instead). Remove escaping from where it made no sense. Removing the code that adds the cache URL, since this is not used nor configured anywhere as yet (also the name of the method was misleading).