Multitenancy support refactoring
Leveraging the WrappedRequest mechanism, thus embedding the OrganizationConfiguration for the currently accessed domain within the request. The MultitenantAction fetches the correct OrganizationConfiguration based on the domain and passes it on to the MultitenantRequest, instead of the previous situation where we were keeping a global map of RequestHeader -> OrganizationConfiguration around. Thus multi-tenancy is now type-safe, i.e. the configuration method available to controllers mixing in OrganizationConfigurationAware trait will only compile if it finds a MultitenantRequest in the implicit scope.

Also starting to simplify action composition, e.g. for the OrganizationAdmin trait it is not anymore necessary to compose against an Action[A] but instead a Request => Result function can directly be passed as body.