Fixing processing for large sets
It appears that MongoDB has very slow writes (108 ms / record locally) with the FsyncSafe WriteConcern. I suppose this is new in the 2.4 series of MongoDB, as it wasn't the case earlier on. This caused a congestion in the processing pipeline which eventually leads to an out-of-memory error. I am now relying on MongoDBs claim that records will eventually get written from memory to disk.

Additionally, there's now a concrete number displayed next to the progress bar.